Place individual subsystem command utils in their own package
This reduces massive kludge-files of doom when we have more complex
logic and makes filenames make sense.
